Also changed test of MDA weights to not use the plugin
