--- devel/test-data/local/custom_kw/iteration_weights/constant_2
@@ -1,2 +1,0 @@
-1.414213562373095
-1.414213562373095
