--- devel/test-data/local/custom_kw/iteration_weights/constant_4
@@ -1,4 +1,0 @@
-2
-2
-2
-2
