--- devel/test-data/local/custom_kw/iteration_weights/error_in_weights
@@ -1,4 +1,0 @@
-2
-2
-error
-2
